feat(#6): BeanMemoryStoreAdvice would allow to define suppliers for empty collections
Today the creation of empty collections is closed within BeanMemoryStoreAdvice. The problem is that when read/write proxy is created and we nee to get empty collection to add a new element to it, Collections.emptyList() / emptyMap() / emptySet() throw an exception since it's read-only.

This approach cannot be easily avoided when using external libraries such as Jackson for filling up the collections so it would be better to adapt BeanMemoryStoreAdvice to accept functions for creating instances of empty collections instead. The code that knows that, this approach will be used can supply their own empty collection instances such as `new LinkedList()`.