fix(#8): Invalid ClassLoader is used for classes created by custom classloader
In order to fix the issue I had to create a private lookup in "context class", which is the class loaded by the target classloader where the proxy class should be loaded. Another problem is with the package access - we need to keep the original package because the created lookup is not able to load class in our "generated" package which is loaded by different classloader.